+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.0.7-alpha.10 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.9 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.8 (2021-01-19)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.7 (2021-01-15)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.6 (2021-01-14)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.5 (2021-01-14)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.4 (2021-01-14)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.3 (2021-01-11)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.2 (2021-01-08)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.1 (2021-01-06)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
+### 1.0.7-alpha.0 (2021-01-05)
+
+**Note:** Version bump only for package react-magma-icons
+
+
+
+
+
 ### 1.0.6 (2020-12-30)
 
 **Note:** Version bump only for package react-magma-icons